+++ src/soopervisor/airflow/export.py
@@ -53,7 +53,7 @@ class AirflowExporter(abc.AbstractExporter):
         pass
 
     @staticmethod
-    def _export(cfg, env_name, mode, until, skip_tests):
+    def _export(cfg, env_name, mode, until, skip_tests, ignore_git):
         """
         Copies the current source code to the target environment folder.
         The code along with the DAG declaration file can be copied to
@@ -74,7 +74,8 @@ class AirflowExporter(abc.AbstractExporter):
                 env_name,
                 until=until,
                 entry_point=args[1],
-                skip_tests=skip_tests)
+                skip_tests=skip_tests,
+                ignore_git=ignore_git)
 
             dag_dict = generate_airflow_spec(tasks, args, target_image)